Revert "Moved usfm package to new 'parse' package and created empty IParser interface"
This reverts commit ddd6f46101c18f893429af7592b9853056747676.
