+++ src/main/kotlin/org/wycliffeassociates/otter/common/domain/resourcecontainer/ResourceContainerExtensions.kt
@@ -1,7 +1,7 @@
 package org.wycliffeassociates.otter.common.domain.resourcecontainer
 
 import org.wycliffeassociates.otter.common.data.model.Collection
-import org.wycliffeassociates.otter.common.domain.parse.usfm.ParseUsfm
+import org.wycliffeassociates.otter.common.domain.usfm.ParseUsfm
 import org.wycliffeassociates.resourcecontainer.ResourceContainer
 import org.wycliffeassociates.resourcecontainer.entity.Project
 import java.io.File
